test-requirements.txt
# Radon requires colorama<0.4, but many other tools listed here will
# install a later version, so when pip tries to install radon it will
# ignore the dependency version difference and only fail when the
# bear is loaded.
colorama<0.4
# coverage 4.4.2 broke compatibility with coverage-config-reload-plugin
# and thus broke https://github.com/jayvdb/coverage_env_plugin .
# See https://github.com/jayvdb/coverage_config_reload_plugin/issues/1
coverage==4.4.1
coverage-env-plugin~=0.1
coverage-config-reload-plugin~=0.2
codecov~=2.0.5
moban~=0.3.3 ; python_version > '3.0'
packaging~=16.8
pytest~=3.6.4
pytest-cov~=2.4
pytest-env~=0.6.0
pytest-instafail~=0.3.0
pytest-mock~=1.1
pytest-profiling~=1.3.0
pytest-reorder~=0.1.0
#git+https://github.com/jayvdb/pytest-reqs@coala#egg=pytest-reqs
pytest-timeout~=1.3.0
pytest-travis-fold~=1.3.0
requests-mock~=1.2
ipdb~=0.11
pip<10
wheel~=0.29
ruamel.yaml~=0.15.42
pytest-pythonpath~=0.7.0
twine~=1.7.4
